requirements.txt
selenium~=4.4.3
pyyaml~=6.0